+++ generate/config.go
@@ -6,8 +6,6 @@ import (
 	"go/token"
 	"os"
 	"path/filepath"
-	"unicode"
-	"unicode/utf8"
 
 	"golang.org/x/tools/go/packages"
 	"gopkg.in/yaml.v2"
@@ -30,7 +28,7 @@ type Config struct {
 	ContextType      string                  `yaml:"context_type"`
 	ClientGetter     string                  `yaml:"client_getter"`
 	Bindings         map[string]*TypeBinding `yaml:"bindings"`
-	AutoBindings     StringList              `yaml:"auto_bindings"`
+	PackageBindings  []*PackageBinding       `yaml:"package_bindings"`
 	StructReferences bool                    `yaml:"use_struct_references"`
 	Extensions       bool                    `yaml:"use_extensions"`
 
@@ -54,6 +52,13 @@ type TypeBinding struct {
 	ExpectExactFields string `yaml:"expect_exact_fields"`
 	Marshaler         string `yaml:"marshaler"`
 	Unmarshaler       string `yaml:"unmarshaler"`
+}
+
+// A PackageBinding represents a Go package for which genqlient will
+// automatically generate TypeBindings, and is documented further at:
+// https://github.com/Khan/genqlient/blob/main/docs/genqlient.yaml
+type PackageBinding struct {
+	Package string `yaml:"package"`
 }
 
 // ValidateAndFillDefaults ensures that the configuration is valid, and fills
@@ -92,12 +97,12 @@ func (c *Config) ValidateAndFillDefaults(baseDir string) error {
 		c.Package = base
 	}
 
-	if len(c.AutoBindings) > 0 {
-		for _, binding := range c.AutoBindings {
+	if len(c.PackageBindings) > 0 {
+		for _, binding := range c.PackageBindings {
 			mode := packages.NeedImports | packages.NeedTypes | packages.NeedTypesSizes
 			pkgs, err := packages.Load(&packages.Config{
 				Mode: mode,
-			}, binding)
+			}, binding.Package)
 			if err != nil {
 				return err
 			}
@@ -109,12 +114,11 @@ func (c *Config) ValidateAndFillDefaults(baseDir string) error {
 			for _, pkg := range pkgs {
 				p := pkg.Types
 				if p == nil || p.Scope() == nil {
-					return errorf(nil, "type not found. autobind %s fail", binding)
+					return errorf(nil, "unable to bind package %s: no types found", binding.Package)
 				}
 
 				for _, typ := range p.Scope().Names() {
-					runeIsUpper, _ := utf8.DecodeRuneInString(typ)
-					if unicode.IsUpper(runeIsUpper) {
+					if token.IsExported(typ) {
 						// Check if type is manual bindings
 						_, exist := c.Bindings[typ]
 						if !exist {